It is working! It is know connected to the graph, it creates and fetches! there is some stuff missing

+++ services/spots/internal/core/domain/event.go
@@ -4,7 +4,7 @@ type Event struct {
 	Name           string `json:"name"`
 	UUID           string `json:"id"`          // Only fields with Cappital letter are exported / Visibles to the outside
 	Description    string `json:"description"` // So as Json use it, we need to make ut public.
-	MaximunCapacty int8   `json:"maximunCapacty"`
+	MaximunCapacty int64  `json:"maximunCapacty"`
 	EventType      string `json:"eventType"`
 }
